+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 **ATTN**: This project uses [semantic versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [v0.5.0] - 2024-05-11
 ### Fixed
 - Added translation to Coordinates block on characters detail view.
 - Fixed destroy cursor in safehouses. Now you cannot destroy stuff inside last lines of Safehouses. To enable this tweak set `FixDestroyCursorInSafehouse` to true in sandbox options.
@@ -88,7 +90,8 @@ All notable changes to this project will be documented in this file.
 - Added renaming light bulbs to improve sorting (English). Cannot be disabled.
 - Added renaming seeds and seeds packets to improve sorting (English). Cannot be disabled.
 
-[Unreleased]: https://github.com/openzomboid/server-tweaker/compare/v0.4.1...HEAD
+[Unreleased]: https://github.com/openzomboid/server-tweaker/compare/v0.5.0...HEAD
+[v0.5.0]: https://github.com/openzomboid/server-tweaker/compare/v0.4.1...v0.5.0
 [v0.4.1]: https://github.com/openzomboid/server-tweaker/compare/v0.4.0...v0.4.1
 [v0.4.0]: https://github.com/openzomboid/server-tweaker/compare/v0.3.2...v0.4.0
 [v0.3.2]: https://github.com/openzomboid/server-tweaker/compare/v0.3.1...v0.3.2